+++ src/main/java/com/finalproject/Tiket/Pesawat/payload/dto/auth/LoginDto.java
@@ -4,6 +4,7 @@ import com.finalproject.Tiket.Pesawat.dto.validation.ValidPassword;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -17,6 +18,6 @@ public class LoginDto {
 
     @NotEmpty(message = "Password is required.")
     @NotBlank(message = "Password cannot be blank.")
-    @ValidPassword
+    @Size(min = 8, message = "Password length min 8 character")
     private String password;
 }
