+++ src/main/java/com/finalproject/Tiket/Pesawat/service/ArrivalService.java
@@ -1,4 +1,4 @@
-package com.finalproject.Tiket.Pesawat.controller;
+package com.finalproject.Tiket.Pesawat.service;
 
 import com.finalproject.Tiket.Pesawat.dto.arrival.response.ResponseListArrival;
